requirements.txt
torch
numpy
nibabel
scikit-image
scikit-learn
tqdm
matplotlib
pyyaml
tensorboard
wandb
pytest
